调整逻辑结构以适应组合内混合股票基金展示

+++ lib/models/PortfolioDetail.dart
@@ -13,7 +13,7 @@ class PortfolioDetail {
   int? over;
   dynamic overIncome;
   double? unitPrice;
-  int? amount;
+  double? amount;
   DateTime? purchaseTime;
   DateTime? updateTime;
   String? location;
